Patched Refined Oil
Math was bad. I fixed it. Not the best fix but it works now. Instead of doing the math to calculate the number of refineries needed, it simply multiplies the needed rate by 4. Using the Reforming Refine recipe (3x refined oil output), the number of refineries needed is always equal to 4x base rate of refined oil.